Fix that closes #145

+++ emg3d/meshes.py
@@ -777,7 +777,7 @@ def good_mg_cell_nr(max_nr=1024, max_prime=5, min_div=3):
 
     """
     # Primes till 20.
-    primes = np.array([2, 3, 5, 7, 11, 13, 17, 19])
+    primes = np.array([2, 3, 5, 7, 11, 13, 17, 19], dtype=np.int64)
 
     # Sanity check; 19 is already ridiculously high.
     if max_prime > primes[-1]:
